feat: dockerised
