+++ _example/ip_addresses_and_http_hosts.yaml
@@ -1,0 +1,48 @@
+polling_interval: 30
+
+webserver:
+  ip: 0.0.0.0
+  port: 8081
+  event_count: 30
+
+alerting:
+  enabled: false
+  server: localhost
+  port: 25
+  recipient: unlucky_admin@yourdomain.com
+  sender: alerting@yourdomain.com
+
+logging:
+  file_path: /tmp/simple_monitor.log
+  level: DEBUG
+  stream_level: INFO
+
+ip:
+  hosts:
+    server1:
+      hostname: server1
+      ip_address: 192.168.1.2
+    server2:
+      hostname: server2
+      ip_address: 192.168.1.3
+    switch1:
+      hostname: switch1
+      ip_address: 192.168.1.4
+    switch2:
+      hostname: switch2
+      ip_address: 192.168.1.5
+    ipcamera1:
+      hostname: ipcamera1
+      ip_address: 192.168.1.16
+    firewall: 
+      hostname: firewall
+      ip_address: 192.168.1.254
+
+http:
+  hosts:
+    kubernetes:
+      hostname: kubernetes.yourdomain.com
+      endpoint: "/healthz"
+    frigate:
+      hostname: netbox.yourdomain.com
+      endpoint: "/healthz"
