+++ backend/requirements.txt
@@ -1,0 +1,7 @@
+certifi==2024.8.30
+charset-normalizer==3.3.2
+idna==3.8
+netaddr==1.3.0
+PyYAML==6.0.2
+requests==2.32.3
+urllib3==2.2.3
