+++ frontend/requirements.txt
@@ -1,0 +1,9 @@
+blinker==1.8.2
+click==8.1.7
+Flask==3.0.3
+itsdangerous==2.2.0
+Jinja2==3.1.4
+MarkupSafe==2.1.5
+PyYAML==6.0.2
+waitress==3.0.0
+Werkzeug==3.0.4
