--- requirements.txt
@@ -1,29 +1,0 @@
-blinker==1.8.2
-certifi==2024.8.30
-charset-normalizer==3.3.2
-click==8.1.7
-colorama==0.4.6
-contourpy==1.3.0
-cycler==0.12.1
-Flask==3.0.3
-fonttools==4.53.1
-idna==3.8
-itsdangerous==2.2.0
-Jinja2==3.1.4
-kiwisolver==1.4.7
-MarkupSafe==2.1.5
-matplotlib==3.9.2
-numpy==2.1.1
-packaging==24.1
-pillow==10.4.0
-psutil==6.0.0
-pyparsing==3.1.4
-python-dateutil==2.9.0.post0
-python-pm==0.0.4
-PyYAML==6.0.2
-requests==2.32.3
-six==1.16.0
-termtables==0.2.4
-urllib3==2.2.3
-waitress==3.0.0
-Werkzeug==3.0.4
